% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getSlicedStats.R
\name{getSlicedStats}
\alias{getSlicedStats}
\title{Calculate multiple statistics for many variables and data subsets}
\usage{
getSlicedStats(data, vars, slice.vars, stats = c("mean", "SE.mean",
  "nbr.val"))
}
\arguments{
\item{data}{This is the data frame that will be operated on}

\item{vars}{Variables to summarize}

\item{slice.vars}{Vector of variables which subset the data set}

\item{stats}{Statistics to calculate. The valid list of values comes from the stat.desc() function}
}
\description{
This function wraps use of data.table(), stat.desc() and helpful data
reshaping courtesy of reshape2.
}
\examples{
getSlicedStats()
}
\keyword{data.table}
\keyword{pastecs}
\keyword{reshape2}
\keyword{stat.desc}

